First commit - spring boot tests eu.mrndesign.matned.repository.ProductRepositoryTest#testProductNotFoundIsCorrect eu.mrndesign.matned.repository.ProductRepositoryTest#testProductCanBeAdded

+++ src/main/java/eu/mrndesign/matned/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package eu.mrndesign.matned.repository;
 import eu.mrndesign.matned.model.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
-    Product findProductById(int i);
+public interface ProductRepository extends CrudRepository<Product, Long> {
+    Product findProductById(Long i);
 }
